requirements.txt
Django==3.1.3
djangorestframework==3.12.2
mysqlclient==2.0.1
django-mysql==3.9
django-cors-headers==3.5.0
mysql-connector-python==8.0.23
mysqlclient==2.0.1
mixer
pytest==6.1.1
pytest-django
python-decouple
pytest-cov
hypothesis
pytest-sugar
pytest-xdist
pytest-black

psycopg2-binary==2.8.4